+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add validation rules for Edit action for the MARC modifications table fields (UIDATIMP-489)
 * Get rid of imported translations for math profiles (UIDATIMP-570)
 * Increment `@folio/stripes` to `v5` and update dependency on `react-router`.
+* Reuse `<EndOfItem>` component from `stripes-data-transfer-components` repository (UIDATIMP-571)
 
 ### Bugs fixed:
 * Fix rendering qualifier sections with old data in match profiles details (UIDATIMP-481)